@@ -11,7 +11,8 @@
   ],
   "require": {
     "guzzlehttp/guzzle": "^6.3",
-    "php":"^7.0"
+    "php":"^7.1",
+    "ext-mbstring": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "6.5.*"
