@@ -6,18 +6,12 @@ namespace Comgate\Response;
 use Comgate\Exception\ErrorCodeException;
 use Comgate\Exception\InvalidArgumentException;
 
-class CreatePaymentResponse extends BaseResponse
+class CreateRepeatedResponse extends BaseResponse
 {
     /**
      * @var string
      */
     private $transId;
-
-    /**
-     * @var string
-     */
-    private $redirect;
-
 
     /**
      * @param array $rawData
@@ -26,18 +20,13 @@ class CreatePaymentResponse extends BaseResponse
      */
     public function __construct(array $rawData)
     {
+
         parent::__construct($rawData);
 
         if (isset($rawData['transId'])) {
             $this->transId = $rawData['transId'];
         } else {
             throw new InvalidArgumentException('Missing "transId" in response');
-        }
-
-        if (isset($rawData['redirect'])) {
-            $this->redirect = $rawData['redirect'];
-        } else {
-            throw new InvalidArgumentException('Missing "redirect" in response');
         }
     }
 
@@ -48,14 +37,5 @@ class CreatePaymentResponse extends BaseResponse
     public function getTransId(): string
     {
         return $this->transId;
-    }
-
-
-    /**
-     * @return string
-     */
-    public function getRedirectUrl(): string
-    {
-        return $this->redirect;
     }
 }
